Start a role for bootstrapping a Kubernetes Cluster on Proxmox
This commit is the beginning of an Ansible playbook for creating a
Kubernetes Cluster on top of a Proxmox Cluster.

So far, there are two roles:
 * ansible-venv, for creating a venv to run Ansible in
 * kubernetes-cluster, a possibly misnamed role for bootstrapping the VMs

The ansible-venv role is required because the proxmox module in
Ansible depends on the `proxmoxer` package from pypi. Normally, the
Debian-packaged version would suffice, but the version in current
stable (bullseye) is too old to support proxmox API tokens. So
installing the version from pip is easier than trying to mess with
repositories on the production nodes.

The kubernetes-cluster role currently just creates the template VMs
and prepares them for cloud-init. A future commit will actually create
the different VMs with their various personalization, which will come
from a locally-executed Ansible playbook.