Initial skeleton of installing kubernetes base packages
This change includes the necessary changes to inventory to be
compatible with ansible-pull, as well as a new playbook
("cluster-bootstrap") that is used by ansible-pull. Currently, this
playbook just installs kubernetes tools, namely kubectl, kubeadm, and
kubelet itself.

The playbook also installs containerd ... from the official releases,
rather than the one packaged in Debian. This is unfortunate, but
kubernetes 1.26 requires the APIs that are only present in containerd
1.6 and newer, which is only available in Debian testing (bookworm) at
this time. There's no backport of containerd either.