Refactor kubernetes-cluster into roles to allow for looping
Previously, the kubernetes-cluster role created the debian VM template
and a kubemaster VM directly. These are not actually specific to the
kubernetes cluster, but generic instantiations of the VM infrastructure.

As a result, they have been split out into 3 roles:

 * proxmox-api, which contains only the defaults and vars for proxmox
   API communications
 * vm-template-debian, which creates the base VM template
 * vm-from-template, which clones the template VM into the desired
   target VM

These can then be instantiated from kubernetes-cluster to create all
the VMs needed. For now, only a single kubemaster and a single worker
are created, but a proper loop can be used to instantiate all of them.

There is a slight wrinkle here -- despite the API being able to
operate on all nodes, the proxmox `qm` command doesn't seem to be able
to work on non-local VMs. So this means that the VMs that run on
different nodes really need to be created directly via Ansible
operating on those hosts.