Modernize proxmox_kvm API a bit
Previously we were using `qm` to do mutating operations, but it turns
out there was a documented `update` knob that I had just missed. So
switch to using that.

Also, it seems that in changing from i440fx to q35, the
start/sleep/stop/start hack no longer is needed, and we can now
reliably cloudinit from a fresh VM. So use the API to start VMs.

While we're here, switch to a slightly newer build of the debian
genericcloud ISO so apt upgrade is faster.