Create a highly available kubernetes cluster
This change adds the haproxy-k8s-api role, which sets up haproxy and
keepalived to proxy to the kubernetes API. When this role is enabled,
the cluster becomes highly-available, because as long as there's at
least one kubemaster running, and at least one haproxy/keepalived pair
running, there will always be somewhere to route the request.

The haproxy-k8s-api role can be used in one of two modes:
 - dedicated-vm, in which 1 (or more, but preferably 2+) VMs are
   installed to run haproxy and keepalived.
 - stacked, in which the existing kubemasters get haproxy and
   keepalived added to their manifests

For most setups, the dedicated-vm option is overkill and
hard-partitions the resources in a not super useful way, so "stacked"
is recommended.

As an added option, the "direct" mode is also supported, in which the
haproxy/keepalived setup will be omitted and instead the API endpoint
of the first kubemaster will be used. This is sufficient for testing
setups, or those where the underlying hypervisor is providing
high-availability for the kubemaster VM.

Finally, this change modifies the kubeadm-init role to actually
provision the other kubemasters and have them join the cluster.