Move cilium installation to kubemasters only
The `cilium install` command only needs to be run once, and must be
run from somewhere with `kubectl` credentials. This means it is not
useful on the kubelets, so don't both downloading it.

Leave the cilium CLI on all the kubemasters, as it also provides some
debugging capabilities. Strictly speaking it's not needed, as it could
always be used remotely, but it's convenient to have it.