Fix up listen addresses to be able to get monitoring  metrics
kubeadm components bind to 127.0.0.1 is a few places, which prevents
prometheus in-cluster from scraping them. Expose these services.

Normally this would be done with a kubeadm config file, but `kubeadm
init` doesn't permit having a config file and also many of the command
line options we specify, so we do it "manually".