Create an initial configuration for Grafana Loki
This change creates an initial configuration for Grafana Loki. This is
surprisingly hard-won, and as a result, not yet fully hooked up to
ArgoCD. This was applied manually with

```
helm upgrade --create-namespace --namespace logging logging .
```

There were a lot of reasons why this was hard-won. For starters, the
Grafana Loki chart claimed there was a `global` section into which you
could put `extraEnvFrom`, but that's not implemented
(https://github.com/grafana/loki/issues/11391).

Next, every part of the installation guide
(https://grafana.com/docs/loki/v2.9.x/setup/install/helm/install-scalable/)
says to just copy-paste your S3 credentials. This is not a good idea
for a variety of reasons, starting with them not being static, as
they're being provisioned by Rook's CephObjectStore integration, and
ending with it just not being good practice to put secrets into
plaintext in a helm chart anyway.

Fortunately, buried in the documentation and the helm chart are clues:
it's possible to use `extraEnvFrom` to load the data from
ConfigMap/Secret, but then you need to pass a special command-line
argument `-config.expand-env=true`. See also, the aforementioned
`global` section that does not actually work.

But, after this adventure, we have centralized logging.

There's still a few TODO issues, like figuring out authentication (see
the sample on https://github.com/grafana/loki/issues/7081 for
guidance). There's also the question of how to automatically configure
the Loki source in Grafana.