Add testing infrastructure for helm charts
It's difficult to check if the vendored helm charts have the right
version in CI, since `helm dependency list` doesn't have a
machine-readable form, and `helm lint` doesn't catch mismatches.

This change adds some infrastructure to parse the output of the helm
CLI (using awk) and a GNU Makefile to be able to run the resulting
tests.